animator/services/voice_synthesis.py
"""
Voice synthesis service using open source TTS models.
Uses Coqui TTS or similar for voice generation.
"""
import os
import uuid
import tempfile
from django.conf import settings


class VoiceSynthesizer:
    """
    Text-to-speech synthesis using open source models.
    Supports multiple voices and languages.
    """

    AVAILABLE_VOICES = {
        'default': {'name': 'Default', 'model': 'tts_models/en/ljspeech/tacotron2-DDC'},
        'male_1': {'name': 'Male Voice 1', 'model': 'tts_models/en/ljspeech/tacotron2-DDC'},
        'female_1': {'name': 'Female Voice 1', 'model': 'tts_models/en/ljspeech/tacotron2-DDC'},
        'narrator': {'name': 'Narrator', 'model': 'tts_models/en/ljspeech/tacotron2-DDC'},
    }

    def __init__(self):
        self.tts = None
        self._init_tts()

    def _init_tts(self):
        """Initialize TTS engine."""
        try:
            from TTS.api import TTS
            # Use a lightweight model by default
            self.tts = TTS(model_name="tts_models/en/ljspeech/tacotron2-DDC")
        except ImportError:
            print("Coqui TTS not available, using fallback")
        except Exception as e:
            print(f"TTS initialization error: {e}")

    def synthesize(self, text: str, voice: str = 'default') -> str:
        """
        Synthesize speech from text.

        Args:
            text: Text to convert to speech
            voice: Voice preset to use

        Returns:
            str: Path to generated audio file
        """
        output_dir = os.path.join(settings.MEDIA_ROOT, 'temp', 'audio')
        os.makedirs(output_dir, exist_ok=True)
        output_path = os.path.join(output_dir, f'voice_{uuid.uuid4()}.wav')

        if self.tts:
            return self._synthesize_with_coqui(text, voice, output_path)
        else:
            return self._synthesize_fallback(text, output_path)

    def _synthesize_with_coqui(self, text: str, voice: str, output_path: str) -> str:
        """Synthesize using Coqui TTS."""
        voice_config = self.AVAILABLE_VOICES.get(voice, self.AVAILABLE_VOICES['default'])

        # Generate speech
        self.tts.tts_to_file(
            text=text,
            file_path=output_path
        )

        return output_path

    def _synthesize_fallback(self, text: str, output_path: str) -> str:
        """
        Fallback synthesis using pyttsx3 or espeak.
        """
        try:
            import pyttsx3

            engine = pyttsx3.init()
            engine.save_to_file(text, output_path)
            engine.runAndWait()

            return output_path
        except ImportError:
            pass

        # Try espeak as last resort
        try:
            import subprocess
            subprocess.run([
                'espeak', '-w', output_path, text
            ], check=True, capture_output=True)
            return output_path
        except:
            pass

        # Generate silent audio as placeholder
        return self._generate_silent_audio(output_path, len(text) / 10)

    def _generate_silent_audio(self, output_path: str, duration: float) -> str:
        """Generate silent audio file as placeholder."""
        try:
            import numpy as np
            from scipy.io import wavfile

            sample_rate = 22050
            num_samples = int(sample_rate * duration)
            silence = np.zeros(num_samples, dtype=np.int16)

            wavfile.write(output_path, sample_rate, silence)
        except ImportError:
            # Create minimal valid WAV file
            import struct

            sample_rate = 22050
            num_samples = int(sample_rate * duration)

            with open(output_path, 'wb') as f:
                # WAV header
                f.write(b'RIFF')
                f.write(struct.pack('<I', 36 + num_samples * 2))
                f.write(b'WAVE')
                f.write(b'fmt ')
                f.write(struct.pack('<I', 16))
                f.write(struct.pack('<H', 1))  # PCM
                f.write(struct.pack('<H', 1))  # Mono
                f.write(struct.pack('<I', sample_rate))
                f.write(struct.pack('<I', sample_rate * 2))
                f.write(struct.pack('<H', 2))
                f.write(struct.pack('<H', 16))
                f.write(b'data')
                f.write(struct.pack('<I', num_samples * 2))
                f.write(b'\x00' * (num_samples * 2))

        return output_path

    def get_available_voices(self) -> dict:
        """Get list of available voices."""
        return {k: v['name'] for k, v in self.AVAILABLE_VOICES.items()}

    def get_voice_preview(self, voice: str) -> str:
        """Generate a preview of a voice."""
        preview_text = "Hello! This is a preview of my voice."
        return self.synthesize(preview_text, voice)
